Denize/29 adding quick donations fragment to bottom navigation bar (#31)
* Sample

* Added fragment for quick donation in navigation bar

* Removed boiler plate code with ButterKnife. Added more dimens sizes, colors, strings into their respective dimens, sizes, & strings XML files.

* Updated dimens XML file with more sizes specified for different UI elements.
