+++ app/src/main/java/com/example/una/MainActivity.java
@@ -11,6 +11,9 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 
 import com.example.una.fragments.HomePageFragment;
+import com.example.una.fragments.ImpactFragment;
+import com.example.una.fragments.MapsViewFragment;
+import com.example.una.fragments.QuickDonateFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 public class MainActivity extends AppCompatActivity {
