+++ app/src/main/java/com/example/una/fragments/MapsViewFragment.java
@@ -1,16 +1,19 @@
-package com.example.una;
+package com.example.una.fragments;
 
 import android.os.Bundle;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-public class QuickDonateFragment extends Fragment {
+import com.example.una.R;
+
+public class MapsViewFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_quick_donate, container, false);
+        return inflater.inflate(R.layout.fragment_map_view, container, false);
     }
 
     @Override
