+++ app/src/main/res/layout/fragment_quick_donate.xml
@@ -7,18 +7,53 @@
     android:layout_height="match_parent">
 
     <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="130dp"
-        android:layout_marginLeft="130dp"
-        android:layout_marginTop="356dp"
-        android:layout_marginEnd="131dp"
-        android:layout_marginRight="131dp"
-        android:layout_marginBottom="356dp"
-        android:text="Quick Donate Fragment"
+        android:id="@+id/tvCharityName"
+        android:layout_width="@dimen/small_textView_width"
+        android:layout_height="@dimen/extra_small_textView_height"
+        android:layout_marginTop="@dimen/small_padding"
+        android:text="TextView"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/ibQuickDonate"
+        android:gravity="center" />
+
+    <TextView
+        android:id="@+id/tvCharityDescription"
+        android:layout_width="@dimen/large_textView_width"
+        android:layout_height="@dimen/extra_small_textView_height"
+        android:layout_marginTop="@dimen/small_padding"
+        android:layout_marginBottom="@dimen/small_padding"
+        android:gravity="center"
+        android:text="TextView"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/tvCharityName"
+        app:layout_constraintVertical_bias="0.0" />
+
+    <ImageButton
+        android:id="@+id/ibQuickDonate"
+        android:layout_width="@dimen/medium_button_width"
+        android:layout_height="@dimen/small_button_height"
+        android:layout_marginTop="@dimen/medium_button_width"
+        android:scaleType="fitCenter"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/tvTitle"
+        app:srcCompat="@drawable/ic_cuppa_24dp"
+        android:background="@color/bumbleYellow"
+        tools:ignore="VectorDrawableCompat" />
+
+    <TextView
+        android:id="@+id/tvTitle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/small_textView_height"
+        android:gravity="center"
+        android:text="@string/text_donate_quickly"
+        android:textSize="40sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
