+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,10 @@
 <resources>
     <dimen name="standard_padding">16dp</dimen>
     <dimen name="small_padding">8dp</dimen>
+    <dimen name="medium_button_width">120dp</dimen>
+    <dimen name="large_textView_width">200dp</dimen>
+    <dimen name="small_textView_width">80dp</dimen>
+    <dimen name="small_button_height">80dp</dimen>
+    <dimen name="small_textView_height">80dp</dimen>
+    <dimen name="extra_small_textView_height">40dp</dimen>
 </resources>
