+++ app/src/main/res/values/strings.xml
@@ -4,5 +4,5 @@
     <string name="text_quick_donate">Frequent</string>
     <string name="text_maps_view">Maps</string>
     <string name="text_impact">Streaks</string>
-
+    <string name="text_donate_quickly">Donate Quickly!</string>
 </resources>
